price updates

+++ src/main/java/com/example/softmeth4/enums/Size.java
@@ -2,16 +2,16 @@ package com.example.softmeth4.enums;
 
 public enum Size {
     SMALL(0),
-    MEDIUM(2),
-    LARGE(4);
+    MEDIUM(2.00),
+    LARGE(4.00);
 
-    private final int price;
+    private final double price;
 
-    Size(int price) {
+    Size(double price) {
         this.price = price;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
     public static Size fromString(String sizeString) {
